requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-HTTPAuth==4.4.0
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
six==1.16.0
Werkzeug==2.0.1
aniso8601==9.0.1
click==7.1.2
Flask==1.1.4
Flask-HTTPAuth==4.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
honcho==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
six==1.16.0
Werkzeug==1.0.1
